requirements.txt
Flask
Flask-Mail
Flask-SQLAlchemy
Flask-Login
email-validator
python-dotenv